Move account to separate package
