+++ geth/account/account.go
@@ -1,4 +1,4 @@
-package node
+package account
 
 import (
 	"bytes"
