+++ geth/api/backend_test.go
@@ -8,6 +8,7 @@ import (
 
 	gethcommon "github.com/ethereum/go-ethereum/common"
 	whisper "github.com/ethereum/go-ethereum/whisper/whisperv5"
+	"github.com/status-im/status-go/geth/account"
 	"github.com/status-im/status-go/geth/api"
 	"github.com/status-im/status-go/geth/common"
 	"github.com/status-im/status-go/geth/jail"
@@ -351,7 +352,7 @@ func (s *BackendTestSuite) TestRaceConditions() {
 			log.Info("AccountManager()")
 			instance := s.backend.AccountManager()
 			s.NotNil(instance)
-			s.IsType(&node.AccountManager{}, instance)
+			s.IsType(&account.AccountManager{}, instance)
 			s.T().Logf("AccountManager(), result: %v", instance)
 			progress <- struct{}{}
 		},
